+++ packages/mosaic/src/chart-builders/chartTypeUtils.ts
@@ -3,14 +3,35 @@ import type {
   ChartBuilderField,
   ChartTypeDefinition,
 } from './types';
-import {buildDefaultChartTitle as buildDefaultChartTitleImpl} from './constants';
 
 export {
   NUMERIC_COLUMN_TYPES,
   TEMPORAL_COLUMN_TYPES,
   QUANTITATIVE_COLUMN_TYPES,
-  buildDefaultChartTitle,
 } from './constants';
+
+/**
+ * Build a default chart title from description and field values
+ */
+export function buildDefaultChartTitle(
+  description: string,
+  fieldValues: Record<string, string>,
+): string {
+  const baseTitle = description.replace(/^Create (a |an )?/, '');
+  const selectedFields = Object.values(fieldValues).filter(Boolean);
+
+  return selectedFields.length > 0
+    ? `${baseTitle} - ${selectedFields.join(', ')}`
+    : baseTitle;
+}
+
+/**
+ * Create a title builder function from a description string
+ */
+export function titleFromDescription(description: string) {
+  return (fieldValues: Record<string, string>) =>
+    buildDefaultChartTitle(description, fieldValues);
+}
 
 export function columnMatchesFieldTypes(
   column: ChartBuilderColumn,
@@ -57,7 +78,7 @@ export function buildChartTypeTitle(
 ): string {
   return chartType.buildTitle
     ? chartType.buildTitle(fieldValues)
-    : buildDefaultChartTitleImpl(chartType.description, fieldValues);
+    : buildDefaultChartTitle(chartType.description, fieldValues);
 }
 
 export function canCreateChartFromType(
