+++ packages/mosaic/src/chart-types/custom-spec/definition.ts
@@ -1,22 +1,19 @@
 import type {Spec} from '@uwdata/mosaic-spec';
 import type {ChartTypeDefinition} from '../base-types';
 import type {CustomSpecChartSettings} from './schema';
-import {buildDefaultChartTitle} from '../../chart-builders/constants';
+import {titleFromDescription} from '../../chart-builders/chartTypeUtils';
 
-function titleFromDescription(description: string) {
-  return (fieldValues: Record<string, string>) =>
-    buildDefaultChartTitle(description, fieldValues);
-}
+const DESCRIPTION = 'Create a chart with custom spec';
 
 export const customSpecChartType: ChartTypeDefinition<CustomSpecChartSettings> =
   {
     id: 'custom-spec',
     label: 'Custom Spec',
-    description: 'Create a chart with custom spec',
+    description: DESCRIPTION,
     aiDescription:
       'Manual template for editing after creation. Prefer explicit chart templates for assistant-created charts.',
     fields: [],
-    buildTitle: titleFromDescription('Create a chart with custom spec'),
+    buildTitle: titleFromDescription(DESCRIPTION),
     createSpec: (tableName): Spec =>
       ({
         plot: [
